WIP moving certain entity models to kindsys
