+++ pkg/codegen/tmpl/entity_summary.tmpl
@@ -1,0 +1,21 @@
+package {{ .MachineName }}
+
+import "github.com/grafana/grafana/pkg/kindsys"
+
+func Summarizer() kindsys.EntitySummaryBuilder
+
+{{ if .BackendCRUDFuncs.Create -}}
+func Create(labels kindsys.LabelSet, obj *{{ .MachineName }}.{{ .Name }}) {
+  return doCreate(labels, obj)
+}
+{{- end }}
+
+{{ if .BackendCRUDFuncs.Get -}}
+func GetByUID(uid string) (*team.Team, error) {
+	return getAllByLabel(labels)
+}
+
+func GetAllByLabel(labels kindsys.LabelSet) []*{{ .MachineName }}.{{ .Name }} {
+  return doCreate(labels, obj)
+}
+{{- end }}
